use mongodb instead of postgres, remove kafka and update dependencies version

+++ internal/application/projection/page.go
@@ -2,10 +2,10 @@ package projection
 
 type (
 	Page[Item any] struct {
-		CurrentPage int32  `json:"currentPage"`
-		PageSize    int32  `json:"pageSize"`
-		TotalPages  int32  `json:"totalPages"`
-		TotalItems  int32  `json:"totalItems"`
+		CurrentPage int64  `json:"currentPage"`
+		PageSize    int64  `json:"pageSize"`
+		TotalPages  int64  `json:"totalPages"`
+		TotalItems  int64  `json:"totalItems"`
 		LastPage    bool   `json:"lastPage"`
 		Items       []Item `json:"items"`
 	}
